cast a long to int in downloader/FileDownloader.java
git-svn-id: https://svn.code.sf.net/p/gsac/code/trunk@1135 60a5d509-8669-43c4-ab2f-0b528a4e2583

+++ src/org/gsac/gsl/downloader/FileDownloader.java
@@ -318,7 +318,7 @@ public class FileDownloader {
                                 continue;
                             }
 
-                            int numBytes = IOUtil.writeTo(
+                            int numBytes = (int)IOUtil.writeTo(
                                                is,
                                                new BufferedOutputStream(
                                                    new FileOutputStream(
